discovery not execution- runInstance problematic
