+++ src/client/testing/testController/common/utils.ts
@@ -1,44 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import * as path from 'path';
-import * as util from 'util';
-import {
-    CancellationToken,
-    Position,
-    Range,
-    TestController,
-    TestItem,
-    TestMessage,
-    TestRun,
-    Uri,
-    Location,
-} from 'vscode';
-import { splitLines } from '../../common/stringUtils';
-import { createDeferred, Deferred } from '../../common/utils/async';
-import { Testing } from '../../common/utils/localize';
-import { traceError, traceVerbose } from '../../logging';
-import { sendTelemetryEvent } from '../../telemetry';
-import { EventName } from '../../telemetry/constants';
-import { TestProvider } from '../types';
-import {
-    clearAllChildren,
-    createErrorTestItem,
-    DebugTestTag,
-    ErrorTestItemOptions,
-    getTestCaseNodes,
-    RunTestTag,
-} from './common/testItemUtilities';
-import {
-    DiscoveredTestItem,
-    DiscoveredTestNode,
-    DiscoveredTestType,
-    ITestDiscoveryAdapter,
-    ITestExecutionAdapter,
-} from './common/types';
-import { fixLogLines } from './common/utils';
-import { IPythonExecutionFactory } from '../../common/process/types';
-
 export function fixLogLines(content: string): string {
     const lines = content.split(/\r?\n/g);
     return `${lines.join('\r\n')}\r\n`;
@@ -87,61 +49,4 @@ export function jsonRPCContent(headers: Map<string, string>, rawData: string): I
         extractedJSON: data,
         remainingRawData,
     };
-}
-function isTestItem(test: DiscoveredTestNode | DiscoveredTestItem): test is DiscoveredTestItem {
-    return test.type_ === 'test';
-}
-
-// had to switch the order of the original parameter since required param cannot follow optional.
-export function populateTestTree(
-    testController: TestController,
-    testTreeData: DiscoveredTestNode,
-    testRoot: TestItem | undefined,
-    wstAdapter: WorkspaceTestAdapter,
-    token?: CancellationToken,
-): void {
-    // If testRoot is undefined, use the info of the root item of testTreeData to create a test item, and append it to the test controller.
-    if (!testRoot) {
-        testRoot = testController.createTestItem(testTreeData.path, testTreeData.name, Uri.file(testTreeData.path));
-
-        testRoot.canResolveChildren = true;
-        testRoot.tags = [RunTestTag, DebugTestTag];
-
-        testController.items.add(testRoot);
-    }
-
-    // Recursively populate the tree with test data.
-    testTreeData.children.forEach((child) => {
-        if (!token?.isCancellationRequested) {
-            if (isTestItem(child)) {
-                const testItem = testController.createTestItem(child.id_, child.name, Uri.file(child.path));
-                testItem.tags = [RunTestTag, DebugTestTag];
-
-                const range = new Range(
-                    new Position(Number(child.lineno) - 1, 0),
-                    new Position(Number(child.lineno), 0),
-                );
-                testItem.canResolveChildren = false;
-                testItem.range = range;
-                testItem.tags = [RunTestTag, DebugTestTag];
-
-                testRoot!.children.add(testItem);
-                // add to our map
-                wstAdapter.runIdToTestItem.set(child.runID, testItem);
-                wstAdapter.runIdToVSid.set(child.runID, child.id_);
-                wstAdapter.vsIdToRunId.set(child.id_, child.runID);
-            } else {
-                let node = testController.items.get(child.path);
-
-                if (!node) {
-                    node = testController.createTestItem(child.id_, child.name, Uri.file(child.path));
-
-                    node.canResolveChildren = true;
-                    node.tags = [RunTestTag, DebugTestTag];
-                    testRoot!.children.add(node);
-                }
-                populateTestTree(testController, child, node, wstAdapter, token);
-            }
-        }
-    });
 }
