+++ src/client/testing/testController/workspaceTestAdapter.ts
@@ -2,19 +2,17 @@
 // Licensed under the MIT License.
 
 import * as util from 'util';
-import { CancellationToken, TestController, TestItem, TestMessage, TestRun, Uri, Location } from 'vscode';
-import { splitLines } from '../../common/stringUtils';
+import { CancellationToken, TestController, TestItem, TestRun, Uri } from 'vscode';
+import path from 'path';
 import { createDeferred, Deferred } from '../../common/utils/async';
 import { Testing } from '../../common/utils/localize';
 import { traceError, traceVerbose } from '../../logging';
 import { sendTelemetryEvent } from '../../telemetry';
 import { EventName } from '../../telemetry/constants';
 import { TestProvider } from '../types';
-import { clearAllChildren, createErrorTestItem, getTestCaseNodes } from './common/testItemUtilities';
+import { ErrorTestItemOptions, createErrorTestItem, getTestCaseNodes } from './common/testItemUtilities';
 import { ITestDiscoveryAdapter, ITestExecutionAdapter } from './common/types';
-import { fixLogLines } from './common/utils';
 import { IPythonExecutionFactory } from '../../common/process/types';
-import { buildErrorNodeOptions } from './common/resultResolver';
 
 /**
  * This class exposes a test-provider-agnostic way of discovering tests.
@@ -62,7 +60,6 @@ export class WorkspaceTestAdapter {
         const deferred = createDeferred<void>();
         this.executing = deferred;
 
-        let rawTestExecData;
         const testCaseNodes: TestItem[] = [];
         const testCaseIds: string[] = [];
         try {
@@ -82,15 +79,16 @@ export class WorkspaceTestAdapter {
 
             // ** execution factory only defined for new rewrite way
             if (executionFactory !== undefined) {
-                rawTestExecData = await this.executionAdapter.runTests(
+                await this.executionAdapter.runTests(
                     this.workspaceUri,
                     testCaseIds,
                     debugBool,
                     executionFactory,
+                    runInstance,
                 );
                 traceVerbose('executionFactory defined');
             } else {
-                rawTestExecData = await this.executionAdapter.runTests(this.workspaceUri, testCaseIds, debugBool);
+                await this.executionAdapter.runTests(this.workspaceUri, testCaseIds, debugBool);
             }
             deferred.resolve();
         } catch (ex) {
@@ -115,147 +113,6 @@ export class WorkspaceTestAdapter {
         } finally {
             this.executing = undefined;
         }
-
-        if (rawTestExecData !== undefined && rawTestExecData.result !== undefined) {
-            // Map which holds the subtest information for each test item.
-            const subTestStats: Map<string, { passed: number; failed: number }> = new Map();
-
-            // iterate through payload and update the UI accordingly.
-            for (const keyTemp of Object.keys(rawTestExecData.result)) {
-                const testCases: TestItem[] = [];
-
-                // grab leaf level test items
-                testController.items.forEach((i) => {
-                    const tempArr: TestItem[] = getTestCaseNodes(i);
-                    testCases.push(...tempArr);
-                });
-
-                if (
-                    rawTestExecData.result[keyTemp].outcome === 'failure' ||
-                    rawTestExecData.result[keyTemp].outcome === 'passed-unexpected'
-                ) {
-                    const rawTraceback = rawTestExecData.result[keyTemp].traceback ?? '';
-                    const traceback = splitLines(rawTraceback, {
-                        trim: false,
-                        removeEmptyEntries: true,
-                    }).join('\r\n');
-
-                    const text = `${rawTestExecData.result[keyTemp].test} failed: ${
-                        rawTestExecData.result[keyTemp].message ?? rawTestExecData.result[keyTemp].outcome
-                    }\r\n${traceback}\r\n`;
-                    const message = new TestMessage(text);
-
-                    // note that keyTemp is a runId for unittest library...
-                    const grabVSid = this.runIdToVSid.get(keyTemp);
-                    // search through freshly built array of testItem to find the failed test and update UI.
-                    testCases.forEach((indiItem) => {
-                        if (indiItem.id === grabVSid) {
-                            if (indiItem.uri && indiItem.range) {
-                                message.location = new Location(indiItem.uri, indiItem.range);
-                                runInstance.failed(indiItem, message);
-                                runInstance.appendOutput(fixLogLines(text));
-                            }
-                        }
-                    });
-                } else if (
-                    rawTestExecData.result[keyTemp].outcome === 'success' ||
-                    rawTestExecData.result[keyTemp].outcome === 'expected-failure'
-                ) {
-                    const grabTestItem = this.runIdToTestItem.get(keyTemp);
-                    const grabVSid = this.runIdToVSid.get(keyTemp);
-                    if (grabTestItem !== undefined) {
-                        testCases.forEach((indiItem) => {
-                            if (indiItem.id === grabVSid) {
-                                if (indiItem.uri && indiItem.range) {
-                                    runInstance.passed(grabTestItem);
-                                    runInstance.appendOutput('Passed here');
-                                }
-                            }
-                        });
-                    }
-                } else if (rawTestExecData.result[keyTemp].outcome === 'skipped') {
-                    const grabTestItem = this.runIdToTestItem.get(keyTemp);
-                    const grabVSid = this.runIdToVSid.get(keyTemp);
-                    if (grabTestItem !== undefined) {
-                        testCases.forEach((indiItem) => {
-                            if (indiItem.id === grabVSid) {
-                                if (indiItem.uri && indiItem.range) {
-                                    runInstance.skipped(grabTestItem);
-                                    runInstance.appendOutput('Skipped here');
-                                }
-                            }
-                        });
-                    }
-                } else if (rawTestExecData.result[keyTemp].outcome === 'subtest-failure') {
-                    // split on " " since the subtest ID has the parent test ID in the first part of the ID.
-                    const parentTestCaseId = keyTemp.split(' ')[0];
-                    const parentTestItem = this.runIdToTestItem.get(parentTestCaseId);
-                    const data = rawTestExecData.result[keyTemp];
-                    // find the subtest's parent test item
-                    if (parentTestItem) {
-                        const subtestStats = subTestStats.get(parentTestCaseId);
-                        if (subtestStats) {
-                            subtestStats.failed += 1;
-                        } else {
-                            subTestStats.set(parentTestCaseId, { failed: 1, passed: 0 });
-                            runInstance.appendOutput(fixLogLines(`${parentTestCaseId} [subtests]:\r\n`));
-                            // clear since subtest items don't persist between runs
-                            clearAllChildren(parentTestItem);
-                        }
-                        const subtestId = keyTemp;
-                        const subTestItem = testController?.createTestItem(subtestId, subtestId);
-                        runInstance.appendOutput(fixLogLines(`${subtestId} Failed\r\n`));
-                        // create a new test item for the subtest
-                        if (subTestItem) {
-                            const traceback = data.traceback ?? '';
-                            const text = `${data.subtest} Failed: ${data.message ?? data.outcome}\r\n${traceback}\r\n`;
-                            runInstance.appendOutput(fixLogLines(text));
-                            parentTestItem.children.add(subTestItem);
-                            runInstance.started(subTestItem);
-                            const message = new TestMessage(rawTestExecData?.result[keyTemp].message ?? '');
-                            if (parentTestItem.uri && parentTestItem.range) {
-                                message.location = new Location(parentTestItem.uri, parentTestItem.range);
-                            }
-                            runInstance.failed(subTestItem, message);
-                        } else {
-                            throw new Error('Unable to create new child node for subtest');
-                        }
-                    } else {
-                        throw new Error('Parent test item not found');
-                    }
-                } else if (rawTestExecData.result[keyTemp].outcome === 'subtest-success') {
-                    // split on " " since the subtest ID has the parent test ID in the first part of the ID.
-                    const parentTestCaseId = keyTemp.split(' ')[0];
-                    const parentTestItem = this.runIdToTestItem.get(parentTestCaseId);
-
-                    // find the subtest's parent test item
-                    if (parentTestItem) {
-                        const subtestStats = subTestStats.get(parentTestCaseId);
-                        if (subtestStats) {
-                            subtestStats.passed += 1;
-                        } else {
-                            subTestStats.set(parentTestCaseId, { failed: 0, passed: 1 });
-                            runInstance.appendOutput(fixLogLines(`${parentTestCaseId} [subtests]:\r\n`));
-                            // clear since subtest items don't persist between runs
-                            clearAllChildren(parentTestItem);
-                        }
-                        const subtestId = keyTemp;
-                        const subTestItem = testController?.createTestItem(subtestId, subtestId);
-                        // create a new test item for the subtest
-                        if (subTestItem) {
-                            parentTestItem.children.add(subTestItem);
-                            runInstance.started(subTestItem);
-                            runInstance.passed(subTestItem);
-                            runInstance.appendOutput(fixLogLines(`${subtestId} Passed\r\n`));
-                        } else {
-                            throw new Error('Unable to create new child node for subtest');
-                        }
-                    } else {
-                        throw new Error('Parent test item not found');
-                    }
-                }
-            }
-        }
         return Promise.resolve();
     }
 
@@ -267,7 +124,7 @@ export class WorkspaceTestAdapter {
         executionFactory?: IPythonExecutionFactory,
     ): Promise<void> {
         sendTelemetryEvent(EventName.UNITTEST_DISCOVERING, undefined, { tool: this.testProvider });
-
+        console.log('dfsad', isMultiroot, workspaceFilePath);
         // Discovery is expensive. If it is already running, use the existing promise.
         if (this.discovering) {
             return this.discovering.promise;
@@ -310,4 +167,13 @@ export class WorkspaceTestAdapter {
         }
         return Promise.resolve();
     }
+}
+
+export function buildErrorNodeOptions(uri: Uri, message: string, testType: string): ErrorTestItemOptions {
+    const labelText = testType === 'pytest' ? 'Pytest Discovery Error' : 'Unittest Discovery Error';
+    return {
+        id: `DiscoveryError:${uri.fsPath}`,
+        label: `${labelText} [${path.basename(uri.fsPath)}]`,
+        error: message,
+    };
 }
